move investment_length_unit_type to separate file
