+++ interest_calculator.py
@@ -2,22 +2,7 @@
 # investment, then display their earnings.
 
 from enum import Enum
-
-class InvestmentLengthUnitType(Enum):
-    MONTHS = ("Months", 30)
-    QUARTERS = ("Quarters", 90)
-    YEARS = ("Years", 365)
-
-    days: int
-    years: float
-
-    def __init__(self, name, number_of_days):
-        self._name_ = name
-        self.days = number_of_days
-        self.years = number_of_days/365
-
-    def __str__(self):
-        return self._name_
+from investment_length_unit_type import InvestmentLengthUnitType
 
 class InvestmentLengthUnitWrapper:
     investment_length_unit_type: InvestmentLengthUnitType
